# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Logistics Enquiry Triage Tool",
    "version": "0.1.0"
  },
  "paths": {
...
    "/actions": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Create Actions",
        "description": "POST /actions \u2014 create action rows for outbound sends (Fix 2, Fix 3).\n\nPrecondition checks (independent of graph state):\n  - analysis.status == 'approved'\n  - verdict is not BLOCK\n  - QUARANTINE must be confirmed (typed_reason provided)\n  - destination must be enabled (tool_config)\n  - breaker must not be open for the destination\n  - expected_analysis_version must match\n  - first send per (enquiry, destination) uses idempotency key enquiry_id:destination\n  - resend requires resend=true and resend_reason",
        "operationId": "create_actions_actions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Actions Actions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "actions"
        ],
        "summary": "List Actions",
        "description": "GET /actions?enquiry_id= \u2014 list actions for an enquiry.",
        "operationId": "list_actions_actions_get",
        "parameters": [
          {
            "name": "enquiry_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Filter by enquiry_id",
              "title": "Enquiry Id"
            },
            "description": "Filter by enquiry_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response List Actions Actions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/actions/{action_id}/retry": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Retry Action",
        "description": "POST /actions/{id}/retry \u2014 retry a failed or unknown action (Fix 8).\n\nFor 'unknown' actions:\n  - First attempts Tool.reconcile() automatically\n  - If reconcile resolves, no confirm needed\n  - If reconcile returns None, requires {confirm: true, typed_reason}\nFor 'failed' actions: plain retry, no confirmation needed.",
        "operationId": "retry_action_actions__action_id__retry_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Retry Action Actions  Action Id  Retry Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}